Sank UserManifestResources to PackageLoading.
