+++ Sources/Workspace/UserToolchain.swift
@@ -25,36 +25,6 @@ private let hostExecutableSuffix = ".exe"
 private let hostExecutableSuffix = ""
 #endif
 
-/// Concrete object for manifest resource provider.
-public struct UserManifestResources: ManifestResourceProvider {
-    public let swiftCompiler: AbsolutePath
-    public let libDir: AbsolutePath
-    public let sdkRoot: AbsolutePath?
-
-    public init(
-        swiftCompiler: AbsolutePath,
-        libDir: AbsolutePath,
-        sdkRoot: AbsolutePath? = nil
-    ) {
-        self.swiftCompiler = swiftCompiler
-        self.libDir = libDir
-        self.sdkRoot = sdkRoot
-    }
-
-    /// Creates the set of manifest resources associated with a `swift` executable.
-    ///
-    /// - Parameters:
-    ///     - swiftExectuable: The absolute path of the associated `swift` executable.
-    public init(swiftExectuable: AbsolutePath) throws {
-        let binDir = swiftExectuable.parentDirectory
-        self.init(
-            swiftCompiler: try UserToolchain.swiftCompilers(
-                forBinDir: binDir,
-                searchPaths: UserToolchain.getSearchPaths()).manifest,
-            libDir: UserToolchain.libDir(forBinDir: binDir))
-    }
-}
-
 // FIXME: This is messy and needs a redesign.
 public final class UserToolchain: Toolchain {
 
@@ -225,10 +195,6 @@ public final class UserToolchain: Toolchain {
             lookup: { UserToolchain.lookup(variable: $0, searchPaths: searchPaths) })
     }
 
-    fileprivate static func libDir(forBinDir binDir: AbsolutePath) -> AbsolutePath {
-        return binDir.parentDirectory.appending(components: "lib", "swift", "pm")
-    }
-
     public init(destination: Destination, environment: [String: String] = Process.env) throws {
         self.destination = destination
         self.processEnvironment = environment
@@ -262,7 +228,7 @@ public final class UserToolchain: Toolchain {
         ] + destination.extraCCFlags
 
         // Compute the path of directory containing the PackageDescription libraries.
-        var pdLibDir = UserToolchain.libDir(forBinDir: binDir)
+        var pdLibDir = UserManifestResources.libDir(forBinDir: binDir)
 
         // Look for an override in the env.
         if let pdLibDirEnvStr = Process.env["SWIFTPM_PD_LIBS"] {
